# swagger.yaml
openapi: 3.0.0
paths:
...
  /notifications:
    get:
      operationId: NotificationsApi_getAllNotifications
      summary: Get all notifications
      description: Returns all notifications.
      parameters: []
      responses:
        '200':
          description: Successful operation. Returns notifications and count.
          headers:
            X-Total-Count:
              description: Count of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDTO'
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: &ref_8
        - notifications
      security:
        - bearer: []
        - bearerAuth: []
  /notifications/new:
    get:
      operationId: NotificationsApi_getNewNotifications
      summary: Get new notifications
      description: Returns new notifications.
      parameters: []
      responses:
        '200':
          description: Successful operation. Returns new notifications.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDTO'
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_8
      security:
        - bearer: []
        - bearerAuth: []
  /notifications/progresses:
    get:
      operationId: NotificationsApi_getProgresses
      summary: Get progresses
      description: Returns progresses.
      parameters: []
      responses:
        '200':
          description: Successful operation. Returns progresses.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProgressDTO'
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_8
      security:
        - bearer: []
        - bearerAuth: []
  /notifications/read/all:
    post:
      operationId: NotificationsApi_readAll
      summary: Read all notifications
      description: Returns new notifications.
      parameters: []
      responses:
        '200':
          description: Successful operation. Returns notifications.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDTO'
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_8
      security:
        - bearer: []
        - bearerAuth: []
  /notifications/delete/{notificationId}:
    delete:
      operationId: NotificationsApi_delete
      summary: Delete notifications up to this point
      description: Returns deleted notifications count.
      parameters:
        - name: notificationId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Successful operation. Returns deleted notifications count.
          content:
            application/json:
              schema:
                type: number
        '401':
          description: Unauthorized.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_8
      security:
        - bearer: []
        - bearerAuth: []
...
  /projects/search:
    post:
      operationId: ProjectsAPI_projectSearch
      summary: Search projects
      description: Search projects by filters
      parameters: [ ]
      requestBody:
        required: true
        description: The question of choosing a methodology
        content:
          application/json:
            schema:
              type: string
            examples:
              q:
                value: >-
                  What methodology can I use for production of electricity using
                  renewable energy technologies?
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDTO'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: &ref_10
        - projects
  /projects/compare/documents:
    post:
      operationId: ProjectsAPI_compareDocuments
      summary: Compare documents.
      description: Compare documents.
      parameters: [ ]
      requestBody:
        required: true
        description: Filters.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterDocumentsDTO'
            examples:
              Filter1:
                value:
                  documentId1: '000000000000000000000001'
                  documentId2: '000000000000000000000002'
              Filter2:
                value:
                  documentIds:
                    - '000000000000000000000001'
                    - '000000000000000000000002'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareDocumentsDTO'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_10
  /projects/properties:
    get:
      operationId: ProjectsAPI_getPolicyProperties
      summary: Get all properties
      description: Get all properties
      parameters: [ ]
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PropertiesDTO'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_10
...
info:
  title: Guardian
  description: >-
    The Guardian is a modular open-source solution that includes best-in-class
    identity management and decentralized ledger technology (DLT) libraries. At
    the heart of the Guardian solution is a sophisticated Policy Workflow Engine
    (PWE) that enables applications to offer a requirements-based tokenization
    implementation.
  version: 2.17.0
  contact:
    name: API developer
    url: https://envisionblockchain.com
    email: info@envisionblockchain.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: /api/v1
    description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InternalServerErrorDTO:
      type: object
      properties:
        code:
          type: number
        message:
          type: string
      required:
        - code
        - message
...
    FilterDocumentsDTO:
      type: object
      properties:
        documentId1:
          type: string
        documentId2:
          type: string
        documentIds:
          type: string
      required:
        - documentId1
        - documentId2
        - documentIds
    CompareDocumentsDTO:
      type: object
      properties:
        documents:
          type: object
        left:
          type: object
        right:
          type: object
        total:
          type: object
      required:
        - documents
        - left
        - right
        - total
...
    NotificationDTO:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum:
            - INFO
            - ERROR
            - WARN
            - SUCCESS
        action:
          type: string
          enum:
            - POLICY_CONFIGURATION
            - POLICY_VIEW
            - POLICIES_PAGE
            - SCHEMAS_PAGE
            - TOKENS_PAGE
            - PROFILE_PAGE
        result:
          type: object
        read:
          type: boolean
        old:
          type: boolean
      required:
        - title
        - message
        - type
        - action
        - result
        - read
        - old
    ProgressDTO:
      type: object
      properties:
        action:
          type: string
        message:
          type: string
        progress:
          type: number
        type:
          type: string
          enum:
            - INFO
            - ERROR
            - WARN
            - SUCCESS
        taskId:
          type: string
      required:
        - action
        - message
        - progress
        - type
        - taskId
    ProjectDTO:
      type: object
      properties:
        id:
          type: string
        policyId:
          type: string
        policyName:
          type: string
        registered:
          type: string
        title:
          type: string
        companyName:
          type: string
        sectoralScope:
          type: string
      required:
        - id
        - policyId
        - policyName
        - registered
        - title
        - companyName
        - sectoralScope
    PropertiesDTO:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        value:
          type: string
      required:
        - id
        - title
        - value
...